requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.8
botocore==1.24.8
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.2
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
jmespath==0.10.0
packaging==21.3
pkg_resources==0.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
wrapt==1.13.3
